+++ docs/_includes/subnav.html
@@ -24,7 +24,7 @@
         </li>
         <li><a href="{% link _docs/events.md %}">Events</a>
             <ul>
-                <li><a href="{% link _docs/events-cloudwatch-event.md %}">CloudWatch Event</a></li>
+                <li><a href="{% link _docs/events-cloudwatch-rule.md %}">CloudWatch Event</a></li>
                 <li><a href="{% link _docs/events-cloudwatch-log.md %}">CloudWatch Log</a></li>
                 <li><a href="{% link _docs/events-dynamodb.md %}">DynamoDB</a></li>
                 <li><a href="{% link _docs/events-iot.md %}">IoT</a></li>
